Made component for slider.
